CourseSell20/pyproject.toml
[tool.poetry]
name = "gameshop"
version = "0.1.0"
description = "Site for sell Game"
authors = ["maratahmetov <ax-marat@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
black = "^23.3.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
psycopg2 = "^2.9.6"
psycopg = "^3.1.9"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
